+++ src/index.html
@@ -5,7 +5,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link href="./style.css">
-    <title>Wbpack Exercise</title>
+    <title>To-Do List</title>
 </head>
 
 <body>
