+++ src/index.js
@@ -1,62 +1,14 @@
-/* eslint-disable max-len */
-
 import './style.css';
+import {
+  addTask, editTask, toggleTask, removeTask,
+} from './modules/crud.js';
 
-// window.localStorage.clear();
-
-let allTasks = JSON.parse(localStorage.getItem('tasks'));
+const allTasks = JSON.parse(localStorage.getItem('tasks'));
 let tasks = [];
-class Task {
-  constructor(index, completed, description) {
-    this.index = index;
-    this.completed = completed;
-    this.description = description;
-  }
-}
-
-const addTask = (description) => {
-  const storedTasks = JSON.parse(localStorage.getItem('tasks'));
-  if (storedTasks === null) {
-    allTasks = [];
-  } else {
-    allTasks = JSON.parse(localStorage.getItem('tasks'));
-  }
-  const newTask = new Task(allTasks.length, false, description);
-  allTasks.push(newTask);
-  localStorage.setItem('tasks', JSON.stringify(allTasks));
-  window.location.reload();
-};
-
-const editTask = (index, description) => {
-  const task = allTasks.find((task) => task.index === index);
-  task.description = description;
-  localStorage.setItem('tasks', JSON.stringify(allTasks));
-};
-
-const toggleTask = (index, completed) => {
-  const task = allTasks.find((task) => task.index === index);
-  task.completed = completed;
-  localStorage.setItem('tasks', JSON.stringify(allTasks));
-};
-
-const removeTask = () => {
-  const completedTasks = [];
-  allTasks.forEach((task) => {
-    if (task.completed === false) {
-      completedTasks.push(task);
-    }
-  });
-  for (let i = 0; i < completedTasks.length; i += 1) {
-    completedTasks[i].index = i;
-  }
-  allTasks = completedTasks;
-  localStorage.setItem('tasks', JSON.stringify(allTasks));
-  window.location.reload();
-};
 
 const paper = document.getElementById('paper');
 
-if (allTasks.length === 0) {
+if (allTasks === null) {
   tasks = [];
   if (paper.childNodes.length < 4) {
     const footer = document.getElementById('remove');
@@ -98,10 +50,6 @@ tasks.forEach((task) => {
   });
   const p = document.createElement('p');
   p.innerHTML = task.description;
-  // console.log(task.description)
-  // let img = document.createElement('img');
-  // img.src = '/assets/move.png';
-  // img.alt = 'move';
   div.appendChild(checkbox);
   div.appendChild(p);
   li.appendChild(div);
@@ -120,17 +68,5 @@ tasks.forEach((task) => {
   li.addEventListener('dblclick', () => {
     li.replaceChild(edit, div);
   });
-  // li.appendChild(img);
   document.getElementById('list').appendChild(li);
 });
-
-// tasks.forEach(task => {
-//   const li =
-//     `<li>
-//       <div style="display: flex; padding-left: 10px;">
-//         <input value=${task.completed} type="checkbox" name="checkbox" id="checkbox${task.index}">
-//         <p>${task.description}</p>
-//       </div>
-//   </li>`
-//   document.getElementById('list').innerHTML += li;
-// });
