+++ src/modules/task.js
@@ -1,0 +1,7 @@
+export default class Task {
+  constructor(index, completed, description) {
+    this.index = index;
+    this.completed = completed;
+    this.description = description;
+  }
+}
